User management

+++ src/app/components/course/course-home/schedule-home.component.ts
@@ -213,29 +213,31 @@ export class ScheduleHomeComponent implements OnInit {
   }
 
   deleteRow(data) {
-    this.isRippleLoad = true;
-    this.apiService.deleteStandard(data.standard_id).subscribe(
-      res => {
-        this.isRippleLoad = false;
-        let data = {
-          type: "success",
-          title: "Success",
-          body: "Deleted Successfully"
-        }
-        this.toastCtrl.popToast(data);
-        this.getAllStandardList();
-      },
-      err => {
-        this.isRippleLoad = false;
-        let data = {
-          type: "error",
-          title: "Error",
-          body: err.error.message
-        }
-        this.toastCtrl.popToast(data);
+    if(confirm('Are you sure you want to delete?')){
+      this.isRippleLoad = true;
+      this.apiService.deleteStandard(data.standard_id).subscribe(
+        res => {
+          this.isRippleLoad = false;
+          let data = {
+            type: "success",
+            title: "Success",
+            body: "Deleted Successfully"
+          }
+          this.toastCtrl.popToast(data);
+          this.getAllStandardList();
+        },
+        err => {
+          this.isRippleLoad = false;
+          let data = {
+            type: "error",
+            title: "Error",
+            body: err.error.message
+          }
+          this.toastCtrl.popToast(data);
 
-      }
-    )
+        }
+      )
+    }
   }
 
   // pagination functions
